fix(webhooks): extract signature helpers so the route exports only POST
